service-urgences-grpc/requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==5.28.3